+++ extend/waibuDb/schema/storage.js
@@ -1,0 +1,14 @@
+async function storage () {
+  return {
+    common: {
+      valueFormatter: {
+        exp: (val, data, { req } = {}) => {
+          const dt = new Date(data.exp)
+          return req.format(dt, 'datetime')
+        }
+      }
+    }
+  }
+}
+
+export default storage